feat(vnext): Add VNext submodule and pipeline script
Added a script that will utilize the VNext instance segmentation
to form the masked images to pass into the E2FGVI module.
Currently, all objects found and segmented from the VNext module
will be removed from the resulting images/video produced from the
VNext module.

The VNext module is now included and utilized as a Git submodule
to this E2FGVI Git repo.