improve(vnext): Improve VNext segmentation results
Now only using the segmentation instances of the "person" class from the VNext
model output for the E2FGVI model masks.

Based on observation, it appears that the more dilation the better when using
the VNext segmentation output for the E2FGVI model masks. The dilation makes
our masks more blob-like and fills out the area around the mask a little more.
This appears to improve the results in the case of the "schoolgirls.mp4" video.